Mixer Update
- Added some additional documentation to the `Mixer` module.
- Passes all tests, formatted, ect...
